requirements.txt
neomodel
django_neomodel
django-crispy-forms
crispy-bootstrap5
djangoajax